requirements.txt
django_bleach
requests
pip install requests
pip install schedule